Refactor & Removed unnecessary files & code
