--- src/test/allFilesTest.js
@@ -1,9 +1,0 @@
-let allFiles = require('./allFilesData');
-
-allFiles.displayData('../filesystem')
-.then(response => {
-        console.log(response);
-})
-
-
-
