--- src/test/editorFileTest.js
@@ -1,9 +1,0 @@
-const editorFiles = require('./editorfiles');
-
-// editorFiles.getAllFiles().then((response) => {
-//   console.log(response);
-// });
-
-editorFiles.removeFromList('cool').then((response) => {
-  console.log(response);
-});
